Appendices/flask/flask/requirements.txt
Flask==3.1.0
Werkzeug>=2.3.0
markupsafe>=2.1.1
Jinja2>=3.1.2
pytest
playwright
pytest-playwright
psycopg2-binary
gunicorn